Rename components, add comments, use loading & error components
